Fixed getMemoryUsage() if > 4 gb mem is used.
svn path=/trunk/libavg/; revision=6711

+++ src/base/OSHelper.cpp
@@ -112,7 +112,7 @@ void setEnv(const string & sName, const string & sVal)
 #endif
 }
 
-unsigned getMemoryUsage()
+size_t getMemoryUsage()
 {
 #ifdef __APPLE__
     kern_return_t rc;
@@ -139,7 +139,7 @@ unsigned getMemoryUsage()
     // See 'man proc' for a documentation of this file's contents.
     std::ifstream f("/proc/self/statm");
     f >> vmsize >> rssize;
-    return rssize*getpagesize();
+    return rssize*(size_t)(getpagesize());
 #endif
 #endif
 }
